Extract pcg interface to a separate object.
